+++ src/utils/logger.rs
@@ -18,7 +18,7 @@ use std::io::{self, Write};
 use std::sync::Arc;
 use std::sync::OnceLock;
 use std::time::Duration;
-use tracing::{error, info, warn};
+use tracing::{error, info, warn, Level};
 
 use crate::utils::{csv::CSVFile, fs::FileMode, github::is_valid_token_file};
 
@@ -147,9 +147,11 @@ pub struct Logger {
 impl Logger {
     /// Creates a new logger and sets it as the global default logger.
     ///
+    /// * debug - If true, the logger will log debug messages as well. Otherwise, only info, warning and error messages will be logged.
+    ///
     /// # Returns
     /// The logger, or an error if the logger could not be created.
-    pub fn new() -> Result<Self> {
+    pub fn new(debug: bool) -> Result<Self> {
         let logger = Self {
             progress: Arc::new(MultiProgress::new()),
         };
@@ -158,11 +160,13 @@ impl Logger {
             progress: Arc::clone(&logger.progress),
         };
 
+        let max_level = if debug { Level::DEBUG } else { Level::INFO };
         let subscriber = tracing_subscriber::fmt()
             .with_writer(writer)
             .with_target(false)
             .without_time()
             .with_level(true)
+            .with_max_level(max_level)
             .finish();
 
         tracing::subscriber::set_global_default(subscriber)?;
@@ -178,10 +182,11 @@ impl Logger {
     ///
     /// # Returns
     /// The result of the task
-    pub fn run_task<T, F>(&self, msg: impl Into<String>, f: F) -> anyhow::Result<T>
-    where
-        F: FnOnce() -> anyhow::Result<T>,
-    {
+    pub fn run_task<T>(
+        &self,
+        msg: impl Into<String>,
+        f: impl FnOnce() -> anyhow::Result<T>,
+    ) -> anyhow::Result<T> {
         let task = TaskLogger::new(self, msg)?;
         let result = f();
 
@@ -213,7 +218,7 @@ static TEST_LOGGER: OnceLock<Logger> = OnceLock::new();
 
 /// Returns a reference to a logger that should be used for testing purposes.
 pub fn test_logger() -> &'static Logger {
-    TEST_LOGGER.get_or_init(|| Logger::new().unwrap())
+    TEST_LOGGER.get_or_init(|| Logger::new(true).unwrap())
 }
 /// Logs if the program will create an output file or overwrite an existing one.
 /// In the latter case, it will also check if the user explicitly asked for it.