+++ src/utils/mod.rs
@@ -21,4 +21,3 @@ pub mod github_api;
 pub mod json;
 pub mod logger;
 pub mod regex;
-pub mod shell_commands;